requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils~=1.0.7
pre-commit      # hooks for applying linters on commit
rich~=13.7.0
pytest~=7.4.4
# sh            # for running bash commands in some tests (linux/macos only)

pyrootutils~=1.0.4
albumentations~=1.3.1
numpy~=1.26.3
pillow~=10.2.0
omegaconf~=2.3.0
tensorflow~=2.15.0
packaging~=23.2
setuptools~=60.2.0